rename pilz_testutils
into pilz_industrial_motion_testutils to be more generic.
This package will not be released as binary and should contain helper
functionality for testing motion generation.
