+++ pilz_industrial_motion_testutils/CMakeLists.txt
@@ -1,5 +1,5 @@
 cmake_minimum_required(VERSION 2.8.3)
-project(pilz_testutils)
+project(pilz_industrial_motion_testutils)
 
 ## Add support for C++11, supported in ROS Kinetic and newer
 add_definitions(-std=c++11)
