+++ pilz_industrial_motion_testutils/include/pilz_industrial_motion_testutils/motion_plan_request_builder.h
@@ -22,7 +22,7 @@
 
 namespace moveit { namespace core{ class RobotState; }}
 
-namespace pilz_testutils {
+namespace pilz_industrial_motion_testutils {
 
 /**
  * @brief The MotionPlanRequestBuilder class builds up a complex motion plan request.
