--- pilz_testutils/CHANGELOG.rst
@@ -1,8 +1,0 @@
-^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-Changelog for package pilz_testutils
-^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-
-Forthcoming
------------
-* xml test data loader.
-* Contributors: Alexander Gutenkunst, Fahri Demirci, Hagen Slusarek, Immanuel Martini, Joachim Schleicher, Kai Hu
