+++ pilz_trajectory_generation/package.xml
@@ -53,7 +53,7 @@
   <test_depend>rostest</test_depend>
   <test_depend>rosunit</test_depend>
   <test_depend>cmake_modules</test_depend>
-  <test_depend>pilz_testutils</test_depend>
+  <test_depend>pilz_industrial_motion_testutils</test_depend>
   <test_depend>prbt_moveit_config</test_depend>
   <test_depend>prbt_support</test_depend>
   <test_depend>prbt_pg70_support</test_depend>
